CC-21893 -- Fail task if there are no tables to capture from Table Filters (#90)
[CC-21893](https://confluentinc.atlassian.net/browse/CC-21893)

Currently, based on the table filters, if there are no tables to be
captured, then the Debezium connectors after coming into RUNNING state
would not perform any processing. There is simple a WARNING log stating

> After applying the include/exclude list filters, no changes will be
captured. Please check your configuration

This is making the users to get confused if there is any issue with the
connector.

With this change, the task will fail with the ConnectException such that
the user will be notified with the error when there are no tables to be
captured with the provided table filters which can be a regex.

[CC-21893]:
https://confluentinc.atlassian.net/browse/CC-21893?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ