Convert Debezium Snapshot & Schema Metrics to compatible for exposing (#91)
[CC-27684](https://confluentinc.atlassian.net/browse/CC-27684) -
Exposing the Debezium Metrics to customers via metrics API. As part of
this ticket, exposing the Snapshot and Schema History metrics which
would provide observability of the connector while in this phase.

As, these metrics are type BOOLEAN and STRING which are by default
emitted by the Debezium. Currently, there is a issue while emitting
these metrics as Otel Collector doesn't support STRING and BOOLEAN type
metrics. Hence, converting these metrics to GAUGE_LONG in order to emit
to Otel Collector and thus expose to customers via Metrics API>

[CC-27684]:
https://confluentinc.atlassian.net/browse/CC-27684?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ