Revert "Remove deprecated 'additional-condition' snapshot config (#5899)"
This reverts commit 39a0874.

Rationale:
We are running Debezium 2.4 in Confluent Cloud and upgrading connectors in-place to Debezium 3.x. Some customer connectors may still be using the deprecated `additional-condition` config. Since this option was silently removed in v3, existing connectors relying on it would unexpectedly perform full table snapshots after upgrade.

To ensure a smoother transition, we're temporarily restoring `additional-condition` support with deprecation warnings, giving customers time to migrate to `snapshot.select.statement.overrides`.

We plan to re-deprecate and remove this in a future version with prior notice.