Recover history only if it exists
This is a follow up to https://github.com/confluentinc/debezium/pull/140.
Although, there is no functional gap in the above PR, but for cases where
schema history doesn't exist, the recovery need not be performed.

At present, the recovery is triggered for every case, but becomes a no-op
for cases like these. We don't want to trigger the recovery for such cases.