Remove support for ts_us and ts_ns (#165)
Between v2 and v3, `ts_us` and `ts_ns` fields were added to support more
granular timestamp information in the change events.
This was done in the tickets --
[DBZ-7107](https://issues.redhat.com/browse/DBZ-7107) and
[DBZ-7880](https://issues.redhat.com/browse/DBZ-7880). Hence this leads
to certain changes in the change event's schema.

For avro, the upgraded schema was backward compatible with the previous
schema (generated from v2 code). But for protobuf and json_sr, it was
not backward compatible.

Revert the support for `ts_us` and `ts_ns` to preserve backward
compatibliity in schema evolution for all data formats.
As of today, nanosecond precision in time is not supported by all the
three DBs and microseconds precision is supported for only some of them.
In future if this is requested, we can think of how can this be solved
in a backward compatible way.

✅ UTs/ITs pass
✅ Manual test

For mysql and postgres, created 3 connectors each. These three
connectors were using three different value converters -- protobuf,
json_sr and avro. All these 6 connectors were failing with schema
compatibility exceptions when debezium connector was in-place upgraded
to v3. But along with this change, they were working fine.