Adds a sleep in the SQL Server IT to ensure the server is ready to serve requests (#177)
The ITs run start when the SQL Server docker image prints the log line
`SQL Server is now ready for client connections`. The first IT sometimes
fails with `login failed for user 'sa'. Reason: An error occurred while
evaluating the password` if run immediately after the log line. Though
the error conveys that the password is incorrect, the rest of the tests
run successfully with the same password indicating that the password, in
actual, is fine.
Based on the build logs from past few failed and successful runs, it has
been found that tests are successful if the SQL server can successfully
start a couple of more things because the first IT tries to make a
connection for example -
`SQL Server Audit is starting the audits`
`SQL Trace ID 1 was started by login "sa"`
`The availability replica manager is starting`
There is a chance that the SQL Server is still starting the internal
components responsible for evaluating the password. Thus, in order to
prevent the SQL Server random tests failure with `login failed for user
'sa'`, I have added a sleep of 3 seconds.

Ran the semaphore pipeline 2-3 times and did not see SQL Server IT
failing with the same error.
```
13:11:27.294 sqlserver
2025-08-22 18:56:27.29 Logon       Login failed for user 'sa'. Reason: An error occurred while evaluating the password. [CLIENT: 172.17.0.1]
13:11:27.378 sqlserver
2025-08-22 18:56:27.37 spid20s     Database 'model_msdb' running the upgrade step from version 943 to version 944.
2025-08-22 13:11:27,299 WARN   BlockingSnapshotIT|||test|  Re-attempting connection request in 2 second(s)   [io.debezium.connector.sqlserver.util.TestHelper]
```

Cherry-Picked-From: 388e28d1a6
Cherry-Picked-By: Bhagyashree <bhagyashree@confluent.io>
Cherry-Picked-At: Fri Mar 13 08:29:41 2026 +0000