CC-35836: Adds IT for log position validation method in SQL Server (#202)
This PR adds integration test for the SQLServer to validate that it is
able to resume from minimumLsn if the stored LSN is no more available.
Disables log position check overall as it was resulting in some IT flaky
failures. Only kept the config on in one IT which tests when_needed
flow.
JIRA: https://confluentinc.atlassian.net/browse/CC-35836