CC-35558: Remove the LsnSeen cache (#205)
What?
----
This PR removes the `lsnSeen` cache, which is not actually required. The
replication slot ensures that the LSNs replayed during both the
searching and streaming phases are always consistent.

Background:
The `lsnSeen` cache was [originally
introduced](https://github.com/debezium/debezium/pull/3904) due to an
incorrect assumption about LSN ordering. The earlier code assumed that
an LSN greater than the `startStreamingLsn` (i.e., the WAL resume
position) could not appear in the streaming phase before reaching
`startStreamingLsn`.

Correction:
In practice, it is possible to encounter such LSNs if the connector
fails to flush the last processed commit LSN back to the Postgres server
while still updating the offsets locally (in offsets topic). In this
scenario, higher LSNs may appear earlier in order than the
`startStreamingLsn`.

Why?
----
Removing this unnecessary cache simplifies the code and eliminates
redundant state management. It also prevents potential memory issues. In
one case, the cache even caused memory issues when the LSN gap between
the first LSN replayed by the slot and the `startStreamingLsn` was very
large.

Consider the scenario illustrated in the diagram below. This case was
previously treated as invalid by the earlier code (prior to the
introduction of the `lsnSeen` cache), but in reality it is entirely
possible and should execute without any failures:

<img width="1747" height="800" alt="Screenshot 2025-10-03 at 6 50 55 PM"
src="https://github.com/user-attachments/assets/17152624-d640-4e7f-ad2d-a7cb462d5337"
/>