CC-37278 Backport changes from DBZ-9549 in v2 (#213)
## TLDR
Backports https://github.com/debezium/debezium/pull/6825 into our own
fork.

## Problem
During Schema history, DDL from each kafka record in the Schema history
topic is parsed and results of parsing are stored in-memory in
`DdlEvents` object. It's upon the caller of `parse` to clear the results
of last parse call before making the call to get correct results. This
is not done in schema history recovery.

## Solution
Rather than implicitly storing the results of parsing and depending upon
callers to clear it. `DdlParser` interface has been changed to not store
and simply return the results of parsing to the caller.

## Manual Test
Image without fix -- `v0.2908.0`
Image with fix -- `v0.2911.0-b525c39b-dirty-akanimesh7`

DB history topic has `1717424` entries and it takes ~108 seconds to
complete recovery

One extra config which was needed was
`schema.history.interna.prefer.ddl` was set to true. If this is not
done, the issue can not be reproduced since the ddl parser code path is
not taken, rather the `tableChanges` field from the database schema
history record is used. The default value of this config is `true` in v1
and `false` in v2. `False` is the preferred value.

### Timelines
17:21-17:23 -- memory before SHR
15:29:41 -- SHR triggered
15:31:30 -- SHR completed
`Notice the memory bump here which stays intact`
15:38:40 -- image with fix deployed 
15:39:13 -- SHR started
15:40:48 -- SHR finished 
`Notice that the same bump was not observed`

<img width="2560" height="1095" alt="image"
src="https://github.com/user-attachments/assets/3757cab1-e6e7-4972-8886-836fbc37e530"
/>